reuse the same interval bar and abstract it as a component to be reused
